pyproject.toml
[tool.poetry]
name = "project-elissa"
version = "0.1.0"
description = "The project to build a recommender for feminin hygiene products scrapped from Amazon.fr"
authors = ["yunan.wang"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
selenium = "^4.16.0"
pytest = "^7.4.3"
pymongo = "^4.6.1"
ipykernel = "^6.27.1"
pandas = "^2.1.4"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = [
    "--import-mode=importlib",
]